chore: organize mutations & add color to fragment
