--- src/graphql/mutations/login.ts
@@ -1,18 +1,0 @@
-import gql from 'graphql-tag'
-
-import {AccountFragment} from '../fragments'
-
-const LOGIN = gql`
-  mutation login($username: String!, $password: String!) {
-    login(input: {username: $username, password: $password}) {
-      token
-      account {
-        ...AccountFragment
-        confirmed
-      }
-    }
-  }
-  ${AccountFragment}
-`
-
-export default LOGIN
