--- src/graphql/mutations/signup.ts
@@ -1,17 +1,0 @@
-import gql from 'graphql-tag'
-
-const SIGNUP = gql`
-  mutation signup($email: String!, $password: String!, $username: String!) {
-    signup(input: {email: $email, password: $password, username: $username}) {
-      token
-      account {
-        id
-        username
-        email
-        role
-      }
-    }
-  }
-`
-
-export default SIGNUP
