+++ src/pages/Singleplayer.tsx
@@ -9,7 +9,7 @@ import TypingArea from '@/components/TypingArea'
 import SingleplayerResults from '@/components/SingleplayerResults'
 import {SingleplayerContainer} from '@/styled/Singleplayer'
 import {TypingState} from '@/types/game'
-import ADD_RESULT from '@/graphql/mutations/addResult'
+import {ADD_RESULT} from '@/graphql/mutations'
 
 const Singleplayer: FC = observer(() => {
   const {GameStore, UserStore} = useStore()
