+++ src/pages/Trial.tsx
@@ -10,7 +10,7 @@ import TypingArea from '@/components/TypingArea'
 import SingleplayerResults from '@/components/SingleplayerResults'
 import {SingleplayerContainer} from '@/styled/Singleplayer'
 import {TypingState} from '@/types/game'
-import {ADD_RESULT_TO_TRIAL} from '@/graphql/mutations/addResult'
+import {ADD_RESULT_TO_TRIAL} from '@/graphql/mutations'
 import {TRIAL} from '@/graphql/queries/trials'
 import Button from '@/styled/Button'
 import {TrialDifficulty, TrialMeta} from '@/styled/Trials'
