+++ src/stores/Game.ts
@@ -3,7 +3,7 @@ import {observable, action, flow} from 'mobx'
 
 import {TypingState} from '@/types/game'
 import {client} from '@/services/Client'
-import {GET_WORD_SET, SEEN} from '@/graphql/mutations/addResult'
+import {GET_WORD_SET, SEEN} from '@/graphql/mutations'
 
 let timeout: any = null
 class GameStore {
